Extract utility function to own file
